% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chunkSummaryC}
\alias{chunkSummaryC}
\title{Chunk Summary via R and C++}
\usage{
chunkSummaryC(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A table is provided, which include the names of each column and the results.
}
\description{
Divide the whole dataset into four chunks and calculate the mean and standard deviation of each chunks. An R function calls the C++.
}
\references{
Hadley Wickham (2015) R package.
}
\seealso{
\code{\link[lng]{chunkSummaryR}}
}
