% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounding.R
\name{confounding}
\alias{confounding}
\title{Identify potential confounders for one predictor of interest}
\usage{
confounding(x, y, dat)
}
\arguments{
\item{x}{a vector of data values that contains the predictor of interest}

\item{y}{a vector of data values that contains outcome varible}

\item{dat}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the predictor of interest, outcome varible and all the covariates}
}
\value{
It returns a list that contains the suggested potential confounders.
}
\description{
Generic function to dectect potential confounders for one predictor of interest no matter the outcome varible and predictor of interest are continuous or categorical.
}
\examples{
library(NHANES)
confounding(NHANES$Weight, NHANES$Diabetes, NHANES)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[stats]{lm}}

\code{\link[stats]{glm}}
}
