% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrib.R
\name{decrib}
\alias{decrib}
\title{create tables providing descriptive statistics}
\usage{
decrib(dat, y = NULL, alternative = c("overall", "total", "bygroup"),
  digit = 2, ...)
}
\arguments{
\item{dat}{a data frame or environment (or object coercible by as.data.frame to a data frame) containing the variables to be summarized}

\item{y}{a numeric vector of outcome variable of interest (default is NULL)}

\item{alternative}{a character string specifying the alternative approaches, must be one of "overall" (default), "total" or "bygroup"}

\item{...}{further arguments to be passed to or from methods}

\item{digits}{the digits of the data values (default is 2)}
}
\value{
Generate descriptive statistical tables for different type of outcome variables of interest without missing value
}
\description{
Generic function to create tables providing descriptive statistics summary for overall and in different groups.
}
\references{
Hlavac Marek (2018). stargazer: Well-Formatted Regression and Summary Statistics Tables. R package version 5.2.1. https://CRAN.R-project.org/package=stargazer
}
\seealso{
\code{\link[stats]{anova}}

\code{\link[stats]{kruskal.test}}
}
