% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lng.R
\docType{package}
\name{lng}
\alias{lng}
\alias{lng-package}
\title{lng: Basic functions for observational studies}
\description{
Generic functions to do a general analysis for observational studies, including to provide descriptive statistical tables, summary tables for univariate regression, suggested confounding variables and summary tables for multivariate regression.
}
\details{
Package: lng

Version:0.1.0

Date:2018-4-12

Depends:R(>=3.2.0)

Imports: multcomp, tidyverse,devtools

Suggests:formatR, knitr, rmarkdown
}
\section{Functions}{

The lng package provides four functions:

decrib: \code{\link[lng]{decrib}}

smruni: \code{\link[lng]{smruni}}

confounding:\code{\link[lng]{confounding}}

smrmul: \code{\link[lng]{smrmul}}
}

\section{lng functions}{

}

\author{
Meng Lan   menglan36@gmail.com,

        Xinye Gui  leafly80@163.com,

        Yuwei Ni   yuweiniCornell@gmail.com
}
