% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesample.R
\docType{data}
\name{packagesample}
\alias{packagesample}
\title{Sample data of adjusted weighting}
\format{A data frame with columns:
\describe{
 \item{Age}{Age in years at screening of study participant. Note: Subjects 80 years or older were recorded as 80}
 \item{Education}{Educational level of study participant Reported for participants aged 20 years or older. One of 8thGrade, 9-11thGrade, HighSchool, SomeCollege, or CollegeGrad}
 \item{MaritalStatus}{Marital status of study participant. Reported for participants aged 20 years or older. One of Married, Widowed, Divorced, Separated, NeverMarried, or LivePartner (living with partner)}
 \item{Weight}{Weight in kg}
 \item{Height}{Standing height in cm. Reported for participants aged 2 years or older}
 \item{BMI}{Body mass index (weight/height2 in kg/m2). Reported for participants aged 2 years or older}
 \item{TotChol}{Total HDL cholesterol in mmol/L. Reported for participants aged 6 years or older}
 \item{Diabetes}{Study participant told by a doctor or health professional that they have diabetes. Reported for participants aged 1 year or older as Yes or No}

}}
\source{
US National Center for Health Statistics (NCHS). Since 1999 approximately 5,000 individuals of all ages are interviewed in their homes every year and complete the health examination component of the survey. The health examination is conducted in a mobile examination centre (MEC). \url{http://www.cdc.gov/nchs/nhanes.htm}
}
\usage{
packagesample
}
\description{
Sample data of adjusted weighting
}
\keyword{datasets}
