% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrmul.R
\name{smrmul}
\alias{smrmul}
\title{Multivariate analysis and summary table}
\usage{
smrmul(x, alternative = c("logistic", "linear"), digit = 3, ...)
}
\arguments{
\item{x}{a regression model which generated by users}

\item{alternative}{a character string specifying the alternative approaches, must be one of "linear" (default), "logistic"}

\item{...}{further arguments to be passed to or from methods}

\item{digits}{the digits of the data values (default is 3)}
}
\value{
For multivariate analysis, it returns a summary table only for significant variavbles based on the type of dependent variables. For continuous Y, the table contains parameters, coefficient, p-value, CI for coefficient; while for categorical variable, the table contains the parameters, Odds ratio, p-value, CI for odds ratio.
}
\description{
Generate summary table proving the multivariate analysis for all variabls and dependent variable from the whole dataset
}
\examples{
a<-rnorm(5,3)
b<-rnorm(5,7)
c<-cbind(a,b)
c<-data.frame(c)

fit1 <- lm(a~., data = c)

smrmul(fit1,alternative = "linear")

smrmul(fit1)


}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
