% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkSummaryR.R
\name{chunkSummaryR}
\alias{chunkSummaryR}
\title{chunkSummaryR}
\usage{
chunkSummaryR(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
Generate summary table proving the multivariate analysis for all variabls and dependent variable from the whole dataset
}
